+++ src/options.js
@@ -5,6 +5,7 @@
 
     const highlightSyntaxElement = document.getElementById('syntax-highlight');
     const highlightOcurrencesElement = document.getElementById('highlight-ocurrences');
+    const ignoreWhitespaceElement = document.getElementById('ignore-whitespace');
     const autocollapsePathsElement = document.getElementById('autocollapse-paths');
     const ignorePathsElement = document.getElementById('ignore-paths');
     const statusElement = document.getElementById('status');
@@ -12,12 +13,14 @@
     function saveOptions() {
         const highlightSyntax = highlightSyntaxElement.checked;
         const highlightOcurrences = highlightOcurrencesElement.checked;
+        const ignoreWhitespace = ignoreWhitespaceElement.checked;
         const autocollapsePaths = autocollapsePathsElement.value.split('\n');
         const ignorePaths = ignorePathsElement.value.split('\n');
 
         chrome.storage.sync.set({
             highlightSyntax,
             highlightOcurrences,
+            ignoreWhitespace,
             autocollapsePaths,
             ignorePaths
         }, () => {
@@ -32,11 +35,13 @@
         chrome.storage.sync.get({
             highlightSyntax: true,
             highlightOcurrences: true,
+            ignoreWhitespace: true,
             autocollapsePaths: ['package-lock.json'],
             ignorePaths: []
         }, options => {
             highlightSyntaxElement.checked = options.highlightSyntax;
             highlightOcurrencesElement.checked = options.highlightOcurrences;
+            ignoreWhitespaceElement.checked = options.ignoreWhitespace;
             autocollapsePathsElement.value = options.autocollapsePaths.join('\n');
             ignorePathsElement.value = options.ignorePaths.join('\n');
         });